requirements.txt
attrs==21.4.0
greenlet==1.1.2
iniconfig==1.1.1
numpy==1.22.0
packaging==21.3
pandas==1.3.5
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.29
toml==0.10.2
